Fix tests for browsers that do not throttle resize event
Safari triggers the resize event for every resize call, which then calls
the throttled recalculateBreakpoints().  When the throttled function is
called multiple times, it condenses to two calls, one leading and one
trailing.

Therefore we should expect X calls in browsers that automatically
throttle resize, or 2X calls in browsers that don't.